+++ test/Raven/Tests/Breadcrumbs/MonologTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-class Raven_Tests_MonologBreadcrumbHandlerTest extends PHPUnit_Framework_TestCase
+class Raven_Tests_MonologBreadcrumbHandlerTest extends \PHPUnit\Framework\TestCase
 {
     protected function getSampleErrorMessage()
     {